major rewrite
Since the courses were previously stored as dictionaries, a major rewrite had to happen. They are now stored as (data)classes, which is much cleaner to write, and their types are now checked! Besides that, some bugs were fixed and some things added/modified, such as the 'next' argument for the open command (opening the next course).